+++ main.go
@@ -1,7 +1,9 @@
 package main
 
 import (
+	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"github.com/efekarakus/termcolor"
 	"github.com/gorilla/mux"
@@ -16,6 +18,7 @@ import (
 	"net"
 	"net/http"
 	"os"
+	"tailscale.com/client/tailscale"
 	"tailscale.com/tsnet"
 	"time"
 )
@@ -94,6 +97,27 @@ func SetupLogging() {
 	}).With().Caller().Logger()
 }
 
+func WaitTSReady(ctx context.Context, lc *tailscale.LocalClient) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	defer cancel()
+	for {
+		if ctx.Err() != nil {
+			break
+		}
+		status, _ := lc.Status(ctx)
+		if status.BackendState == "Running" {
+			return nil
+		} else if status.BackendState == "NeedsLogin" {
+			return errors.New("Authentication (Login) required.")
+		} else if status.BackendState == "NeedsMachineAuth" {
+			return errors.New("Authentication (MachineAuth) required.")
+		}
+		time.Sleep(100 * time.Millisecond)
+	}
+
+	return errors.New("timeout reached")
+}
+
 func startTSListener(r *mux.Router, c types.ListenerConfig) error {
 	s := tsnet.Server{
 		Logf:       func(string, ...any) {},
@@ -101,12 +125,24 @@ func startTSListener(r *mux.Router, c types.ListenerConfig) error {
 		ControlURL: c.TS_ControllURL,
 	}
 	defer s.Close()
+	lc, err := s.LocalClient()
+	if err != nil {
+		log.Error().Err(err).Msg("Could not get local client")
+	}
+
+	err = WaitTSReady(context.Background(), lc)
+	if err != nil {
+		log.Error().Err(err).Msg("TS not ready.")
+		return err
+	}
 
 	ln, err := s.Listen("tcp", fmt.Sprintf(":%s", c.Port))
 	if err != nil {
 		log.Error().Err(err)
 	}
 	defer ln.Close()
+
+	log.Debug().Msg("TS ready and listening")
 
 	return http.Serve(ln, r)
 }